refactor(msg-gen): create more helpers
